[ADD] The `Maybe` type
